+++ Automate/Framework/Machines/TerrainFeatures/BushMachine.cs
@@ -10,6 +10,19 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
     internal class BushMachine : BaseMachine<Bush>
     {
         /*********
+        ** Fields
+        *********/
+        /// <summary>The season for which <see cref="BushMachine.InSeasonImpl"/> applies.</summary>
+        private string LastSeason;
+
+        /// <summary>The day of month for which <see cref="BushMachine.InSeasonImpl"/> applies.</summary>
+        private int LastDay;
+
+        /// <summary>Whether the bush was in-season as of the last check.</summary>
+        private bool InSeasonImpl;
+
+
+        /*********
         ** Public methods
         *********/
         /// <summary>Construct an instance.</summary>
@@ -28,18 +41,19 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
             // berry bush
             int itemId = Game1.currentSeason == "fall" ? 410 : 296; // blackberry or salmonberry
             int quality = Game1.player.professions.Contains(Farmer.botanist) ? SObject.bestQuality : SObject.lowQuality;
-            return new TrackedItem(new SObject(itemId, 1, quality: quality), onReduced: this.OnOutputReduced);
+            int count = 1 + Game1.player.ForagingLevel / 4;
+            return new TrackedItem(new SObject(itemId, initialStack: count, quality: quality), onReduced: this.OnOutputReduced);
         }
 
         /// <summary>Get the machine's processing state.</summary>
         public override MachineState GetState()
         {
-            if (this.Machine.tileSheetOffset.Value == 1)
-                return MachineState.Done;
+            if (!this.InSeason())
+                return MachineState.Disabled;
 
-            return this.Machine.inBloom(Game1.currentSeason, Game1.dayOfMonth)
-                ? MachineState.Processing
-                : MachineState.Disabled;
+            return this.Machine.tileSheetOffset.Value == 1
+                ? MachineState.Done
+                : MachineState.Processing;
         }
 
         /// <summary>Provide input to the machine.</summary>
@@ -82,6 +96,19 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.TerrainFeatures
                 width: box.Width / Game1.tileSize,
                 height: box.Height / Game1.tileSize
             );
+        }
+
+        /// <summary>Get whether the bush is currently in-season to produce berries or tea leaves.</summary>
+        private bool InSeason()
+        {
+            if (this.LastSeason != Game1.currentSeason || this.LastDay != Game1.dayOfMonth)
+            {
+                this.InSeasonImpl = this.Machine.inBloom(Game1.currentSeason, Game1.dayOfMonth);
+                this.LastSeason = Game1.currentSeason;
+                this.LastDay = Game1.dayOfMonth;
+            }
+
+            return this.InSeasonImpl;
         }
     }
 }