+++ ContentPatcher/Framework/Patches/IPatch.cs
@@ -11,7 +11,7 @@ namespace ContentPatcher.Framework.Patches
         ** Accessors
         *********/
         /// <summary>The path to the patch from the root content file.</summary>
-        public LogPathBuilder Path { get; }
+        LogPathBuilder Path { get; }
 
         /// <summary>The parent patch for which this patch was loaded, if any.</summary>
         IPatch ParentPatch { get; }