+++ ContentPatcher/release-notes.md
@@ -4,6 +4,7 @@
 ## Upcoming release
 * Fixed patches not always unapplied when an `Include` patch changes.
 * Fixed error using some tokens within the `contains` input argument.
+* Fixed broken error message when multiple load patches apply in 1.16.
 
 ## 1.16.1
 Released 03 August 2020.